Split up bcc tests into multiple files
